Math - Rect migration

+++ Math/v2/_interface/VectorC_def.h
@@ -43,6 +43,6 @@ typedef struct VECTOR2I_TYP
 	union
 	{
 		struct { int x, y; };
-		double Points[2];
+		int Points[2];
 	};
 } VECTOR2I, * VECTOR2I_PTR;
